file refactor fixes
