+++ babel.config.js
@@ -1,6 +1,16 @@
 module.exports = {
-  presets: [
-    'module:metro-react-native-babel-preset',
-    'module:react-native-dotenv',
+  presets: ['module:metro-react-native-babel-preset'],
+  plugins: [
+    [
+      'module:react-native-dotenv',
+      {
+        moduleName: '@env',
+        path: '.env',
+        blacklist: null,
+        whitelist: null,
+        safe: false,
+        allowUndefined: true,
+      },
+    ],
   ],
 };
