--- src/navigations/index.js
@@ -1,1 +1,0 @@
-// https://cheesecakelabs.com/blog/efficient-way-structure-react-native-projects/
